requirements.txt
altair==4.1.0
appnope @ file:///opt/concourse/worker/volumes/live/0291c9e1-4b15-459f-623e-2770f55be269/volume/appnope_1594338395037/work
argon2-cffi @ file:///opt/concourse/worker/volumes/live/59af29ac-4890-416e-7ab7-794f8d6f7ecd/volume/argon2-cffi_1596828548321/work
astor==0.8.1
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2020.6.20
cffi @ file:///opt/concourse/worker/volumes/live/b9607b09-b777-4ff7-53dc-287727eb8574/volume/cffi_1600699191154/work
charset-normalizer==2.0.4
click==7.1.2
cymem==2.0.5
decorator==4.4.2
defusedxml==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
entrypoints==0.3
environment-kernels==1.1.1
gitdb==4.0.7
GitPython==3.1.18
htbuilder==0.5.0
idna==3.2
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///opt/concourse/worker/volumes/live/88f541d3-5a27-498f-7391-f2e50ca36560/volume/ipykernel_1596206680118/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/90527372-a5f7-4d48-68c2-a99de2ba31c7/volume/ipython_1599056199422/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
iteration-utilities==0.11.0
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1610146833874/work
Jinja2==2.11.2
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
MarkupSafe @ file:///opt/concourse/worker/volumes/live/cb778296-98db-45ad-411e-6f726e102dc3/volume/markupsafe_1594371638608/work
mistune @ file:///opt/concourse/worker/volumes/live/95802d64-d39c-491b-74ce-b9326880ca54/volume/mistune_1594373201816/work
murmurhash==1.0.5
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///opt/concourse/worker/volumes/live/2b9c1d93-d0fd-432f-7d93-66c93d81b614/volume/nbconvert_1601914875037/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
notebook @ file:///opt/concourse/worker/volumes/live/be0f3504-189d-4bae-4e57-c5d6da73ffcd/volume/notebook_1601501605350/work
numpy==1.21.2
packaging==20.4
pandas==1.3.2
pandocfilters==1.4.2
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pathy==0.6.0
pexpect @ file:///opt/concourse/worker/volumes/live/8701bb20-ad87-46c7-5108-30c178cf97e5/volume/pexpect_1594383388344/work
pickleshare @ file:///opt/concourse/worker/volumes/live/93ec39d8-05bb-4f84-7efc-98735bc39b70/volume/pickleshare_1594384101884/work
Pillow==8.3.1
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.17.3
ptyprocess==0.6.0
pyarrow==5.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.8.2
pydeck==0.7.0
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyparsing==2.4.7
pyrsistent @ file:///opt/concourse/worker/volumes/live/ff11f3f0-615b-4508-471d-4d9f19fa6657/volume/pyrsistent_1600141727281/work
python-dateutil==2.8.1
pytz==2021.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
requests==2.26.0
Send2Trash==1.5.0
six==1.15.0
smart-open==5.2.1
smmap==4.0.0
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
st-annotated-text==2.0.0
streamlit==0.87.0
terminado==0.9.1
testpath==0.4.4
thinc==8.0.8
toml==0.10.2
toolz==0.11.1
tornado==6.0.4
tqdm==4.62.2
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typer==0.3.2
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.5
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp @ file:///tmp/build/80754af9/zipp_1603306325174/work
